+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This is a **major rewrite** of the library internals, with the following goals i
 
 - The `sx` styling properties are no longer supported; use `style`, `className`, `styles` and `classNames` instead
 - The column `textAlignment` property was renamed to `textAlign`
+- The table `verticalAlignment` property was renamed to `verticalAlign`
 - The internal context-menu functionality was removed in favor of using the [Mantine ContextMenu](https://icflorescu.github.io/mantine-contextmenu/) package (built by the same author), which is more flexible and feature-rich
 - Some method signatures were changed from `(record, index) => ...` to `({ record, index }) => ...` for clarity